Updated to .Net Core 3.0 Preview 2
Used (server-side) Razor Components, as shipped in ASP.Net Core 3.0 Preview 2. This uses BlazorMVVM.Server and BlazorMVVM.App formats, which is required at the moment for proper treatment of .cshtml Razor Components. According to the ASP.Net team, the plan is to rename .cshtml files to .razor for Razor Components, at which point, this can be consolidated back to one project.